use common emit_finished
